# specs/apps/a-demo/contracts/generated/openapi-bundled.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OrganicLever Demo API",
    "description": "REST API contract for the OrganicLever expense tracker demo application",
    "version": "1.0.0",
    "contact": {
      "name": "Open Sharia Enterprise",
      "url": "https://github.com/wahidyankf/open-sharia-enterprise"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Default backend (Go/Gin)"
    },
    {
      "url": "http://localhost:8201",
      "description": "Java Spring Boot backend"
    }
  ],
...
  "paths": {
...
    "/api/v1/expenses": {
      "post": {
        "operationId": "createExpense",
        "summary": "Create a new expense or income record",
        "tags": ["Expenses"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExpenseRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Expense created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Expense"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listExpenses",
        "summary": "List expenses with pagination",
        "tags": ["Expenses"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "currency",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["income", "expense"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated expense list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/expenses/{id}": {
      "get": {
        "operationId": "getExpense",
        "summary": "Get expense by ID",
        "tags": ["Expenses"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Expense details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Expense"
                }
              }
            }
          },
          "404": {
            "description": "Expense not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateExpense",
        "summary": "Update expense by ID",
        "tags": ["Expenses"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExpenseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expense updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Expense"
                }
              }
            }
          },
          "404": {
            "description": "Expense not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteExpense",
        "summary": "Delete expense by ID",
        "tags": ["Expenses"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Expense deleted"
          },
          "404": {
            "description": "Expense not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/expenses/summary": {
      "get": {
        "operationId": "getExpenseSummary",
        "summary": "Get expense summary grouped by currency",
        "tags": ["Expenses"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Expense summary by currency",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExpenseSummary"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/test/promote-admin": {
      "post": {
        "operationId": "promoteToAdmin",
        "summary": "Promote a user to admin role",
        "tags": ["Test Support"],
        "x-test-only": true,
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User promoted to admin"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Expense": {
        "type": "object",
        "description": "An expense or income record",
        "required": [
          "id",
          "amount",
          "currency",
          "category",
          "description",
          "date",
          "type",
          "userId",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "amount": {
            "type": "string",
            "description": "Decimal amount as string for precision"
          },
          "currency": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "type": {
            "type": "string",
            "enum": ["income", "expense"]
          },
          "quantity": {
            "type": "number",
            "nullable": true
          },
          "unit": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateExpenseRequest": {
        "type": "object",
        "description": "Create a new expense or income record",
        "required": ["amount", "currency", "category", "description", "date", "type"],
        "properties": {
          "amount": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "type": {
            "type": "string",
            "enum": ["income", "expense"]
          },
          "quantity": {
            "type": "number",
            "nullable": true
          },
          "unit": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "UpdateExpenseRequest": {
        "type": "object",
        "description": "Update an existing expense or income record (all fields optional)",
        "properties": {
          "amount": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "type": {
            "type": "string",
            "enum": ["income", "expense"]
          },
          "quantity": {
            "type": "number",
            "nullable": true
          },
          "unit": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ExpenseListResponse": {
        "type": "object",
        "description": "Paginated list of expenses",
        "required": ["content", "totalElements", "totalPages", "page", "size"],
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Expense"
            }
          },
          "totalElements": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "size": {
            "type": "integer"
          }
        }
      },
      "CategoryBreakdown": {
        "type": "object",
        "description": "Breakdown of totals by category",
        "required": ["category", "type", "total"],
        "properties": {
          "category": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "total": {
            "type": "string"
          }
        }
      },
      "ExpenseSummary": {
        "type": "object",
        "description": "Summary of expenses by currency",
        "required": ["currency", "totalIncome", "totalExpense", "net", "categories"],
        "properties": {
          "currency": {
            "type": "string"
          },
          "totalIncome": {
            "type": "string"
          },
          "totalExpense": {
            "type": "string"
          },
          "net": {
            "type": "string"
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryBreakdown"
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standardized error response",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
    }
  }
}